Fix cursor flashing to screen bottom issue
sometimes when holding down the arrow key or backspace button, the cursor would flash to the
bottom of the screen.  This was because we were repainting the screen before setting the cursor
position.

To fix this, make the Cursor object a member of Buffer and set the cursor position before
repainting the screen.