Optimize title refresh
When drawing the title window, I put empty spaces in every character and then went back to put
the title string in the window.

This change optimizes that by putting the title string in the window first and then puts empty
spaces in starting at the end of the title string.  No doubling back essentially